renaming Setup to AASetup
