+++ client/src/components/MultiJurisdictionAudit/StatusBox/index.test.tsx
@@ -10,7 +10,7 @@ import {
   roundMocks,
   auditBoardMocks,
 } from '../_mocks'
-import { contestMocks } from '../Setup/Contests/_mocks'
+import { contestMocks } from '../AASetup/Contests/_mocks'
 import * as utilities from '../../utilities'
 
 jest.mock('react-router-dom', () => ({
