+++ client/src/components/MultiJurisdictionAudit/index.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import { ElementType } from '../../types'
 import { Wrapper, Inner } from '../Atoms/Wrapper'
 import Sidebar from '../Atoms/Sidebar'
-import Setup, { setupStages } from './Setup'
+import Setup, { setupStages } from './AASetup'
 import Progress from './Progress'
 import useSetupMenuItems from './useSetupMenuItems'
 import RoundManagement from './RoundManagement'
